requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
click==8.0.1
colorama==0.4.4
fastapi==0.68.0
greenlet==1.1.1
h11==0.12.0
httptools==0.2.0
iniconfig==1.1.1
isort==5.9.3
mysql-connector-python==8.0.26
packaging==21.0
passlib==1.7.4
pluggy==0.13.1
protobuf==3.17.3
py==1.10.0
pycodestyle==2.7.0
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.19.0
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.22
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.15.0
watchgod==0.7
websockets==9.1
